Moved statements under cdk
