+++ backend/dataall/modules/common/sagemaker/cdk/__init__.py
@@ -1,0 +1,1 @@
+# Contains infrastructure code shared between ML studio and notebooks
